feature: add pet picture

+++ app/Http/Controllers/PetController.php
@@ -6,6 +6,8 @@ use App\Http\Requests\PetUpdateRequest;
 use App\Models\Pet;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
 
 class PetController extends Controller
@@ -33,18 +35,28 @@ class PetController extends Controller
      */
     public function store(PetUpdateRequest $request): RedirectResponse
     {
-        $validated = $request->validated();
-        $request->user()->pets()->create($validated);
+        $user = Auth::user();
 
-        return redirect()->route('dashboard')->with('status', 'Pet profile created :)');
+        $validated = $request->validated();
+
+        if ($request->hasFile('picture')) {
+            $path = $request->file('picture')->store('pets', 'public');
+            $validated['picture'] = $path;
+        }
+
+        $user->pets()->create($validated);
+
+        return redirect()->back()->with('status', 'Pet profile created :)');
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Pet $pet)
+    public function show(Pet $pet): View
     {
-        //
+        return view('pets.show', [
+            'pet' => $pet,
+        ]);
     }
 
     /**
@@ -68,10 +80,9 @@ class PetController extends Controller
      */
     public function destroy(Pet $pet): RedirectResponse
     {
-
-//        $this->authorize('delete', $pet);
+        $this->authorize('delete', $pet);
         $pet->delete();
-        return redirect(route('pets.index'))->with('status', 'Pet profile deleted :(');
-//        return redirect()->route('dashboard')->with('status', 'Pet profile created :)');
+        return redirect()->back()->with('status', 'Pet profile deleted :(');
     }
+
 }
